feat: terms객체 파일분리, 텍스트 컴포넌트 memo최적화, RermsModal lazy import 최적화
