+++ src/constants/termsServiceObject.ts
@@ -1,0 +1,381 @@
+const termsServiceObject = {
+  sections: [
+    {
+      id: "1",
+      title: "제1장 총칙",
+      content: "",
+    },
+    {
+      id: "1-1",
+      title: "제1조 목적",
+      content:
+        '이 약관은 "플레이하이브"(이하 "회사")가 서비스하는 PLATHIVE(플레이하이브, 이하 \'서비스\')의 이용조건 및 절차에 관한 사항과 기타 필요사항을 규정함을 목적으로 합니다.',
+    },
+    {
+      id: "1-2",
+      title: "제2조 약관의 공지 및 효력",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "본 약관의 내용은 서비스 화면에 게시하거나 기타의 방법으로 회원에게 공지함으로써 효력이 발생합니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 회원을 위하여 새로운 서비스를 추가하거나 회사 정책상 중요 사유가 있을 경우 약관을 변경할 수 있으며, 이에 대하여 전항과 같은 방법에 의해 효력을 갖습니다.",
+        },
+      ],
+    },
+    {
+      id: "1-3",
+      title: "제3조 약관의 준칙",
+      content:
+        "이 약관에 명시되지 않은 사항은 전기통신기본법, 전기통신사업법 및 기타 관계법령의 규정에 의합니다.",
+    },
+    {
+      id: "2",
+      title: "제2장 서비스 이용계약",
+      content: "",
+    },
+    {
+      id: "2-1",
+      title: "제4조 이용계약의 성립",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            '이용자가 이용신청시의 "동의함" 단추를 누름과 동시에 이 약관에 동의하는 것으로 간주됩니다.',
+        },
+        {
+          id: "b",
+          content:
+            "이용계약은 이용자의 이용신청에 대하여 회사가 승낙함으로써 성립합니다.",
+        },
+        {
+          id: "c",
+          content:
+            "회원에 가입하고자 하는 이용자는 회사에서 요청하는 개인정보 및 기타 부가정보를 제공해야 합니다.",
+        },
+        {
+          id: "d",
+          content:
+            "이용계약은 회원 ID당 1개의 ID로 제한하는 것을 원칙으로 합니다.",
+        },
+      ],
+    },
+    {
+      id: "2-2",
+      title: "제5조 이용신청의 승낙 및 제한",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 회원(또는 법인)이 이용신청에 정한 모든 사항을 정확히 기재하여 이용신청을 하였을 때 승낙합니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 다음에 해당하는 경우에 승낙을 거부하거나 등록 후라도 회원에게 통보하지 않고 회원정보를 수정 또는 삭제할 수 있습니다.",
+        },
+      ],
+      circledItems: [
+        {
+          id: "1",
+          content: "다른 사람의 명의를 사용하여 신청하였을 경우",
+        },
+        {
+          id: "2",
+          content: "이용계약 신청서의 내용을 허위로 기재하였을 경우",
+        },
+        {
+          id: "3",
+          content:
+            "사회의 안녕질서 및 미풍양속을 저해할 목적으로 신청하였을 경우",
+        },
+        {
+          id: "4",
+          content: "기타 회사가 정한 이용신청 요건이 미비하였을 경우",
+        },
+        {
+          id: "5",
+          content: "기업 신청자가 만 14세 미만인 경우",
+        },
+      ],
+    },
+    ,
+    {
+      id: "2-3",
+      title: "제6조 개인정보의 수집",
+      content:
+        "회원은 본 서비스의 이용을 위해 자신의 정보를 성실히 관리해야 하며 개인정보의 변동사항이 있을 경우 '내정보수정' 페이지에서 이를 수정해야 합니다.",
+    },
+    {
+      id: "2-4",
+      title: "제7조 회원정보의 공유",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사가 타 업체와 제휴, 인수, 분사, 합병 시 회원의 정보는 공유될 수 있습니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회원은 회사에 제공한 개인정보의 수집과 이용에 대한 동의를 언제든지 철회할 수 있습니다.",
+        },
+      ],
+    },
+    {
+      id: "2-5",
+      title: "제8조 개인정보의 보호",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 이용자의 개인정보를 수집할 때 반드시 본인의 동의를 받습니다.",
+        },
+      ],
+    },
+    {
+      id: "3",
+      title: "제3장 서비스 제공 및 변경, 게시물에 대한 저작권",
+      content: "",
+    },
+    {
+      id: "3-1",
+      title: "제9조 서비스 이용",
+      content:
+        "회사는 하루 24시간 365일 서비스 제공을 원칙으로 합니다. 단, 정기점검이나 설비의 보수, 전기통신 사업법에 의한 기간통신사업자의 서비스 중단 등의 이유로 사전공지 후 혹은 천재지변 등 불가항력적인 사유로 사전공지 없이 서비스가 일시 중단될 수 있습니다.",
+    },
+    {
+      id: "3-2",
+      title: "제10조 게시물, 내용물 및 회원정보의 삭제",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 서비스의 게시물 또는 내용물이 제12조의 규정에 위반되는 경우 사전통지 없이 삭제할 수 있습니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 상법의 공항에 의거가 없다고  판단되면 필요에 따라 이용자의 파일정보 및 회원정보를 사전공지 후 삭제할 수 있습니다.",
+        },
+      ],
+    },
+    {
+      id: "3-3",
+      title: "제11조 회원의 게시물에 대한 저작권 등",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content: "회원의 게시물에 관한 규정은 다음과 같습니다.",
+        },
+      ],
+      circledItems: [
+        {
+          id: "1",
+          content:
+            "게시물이라 함은 회사에서 제공하는 모든 게시판과 댓글에 회원이 올린 글, 사진, 각종 파일과 링크 등을 포함합니다.",
+        },
+        {
+          id: "2",
+          content:
+            "회원이 게시하는 정보 등으로 인해 발생하는 손실이나 문제는 전적으로 회원 개인의 판단에 따른 책임이며, 회사의 고의가 아닌 한, 회사는 이에 대하여 책임지지 않습니다.",
+        },
+        {
+          id: "3",
+          content:
+            "회원은 공공질서나 미풍양속에 위배되는 내용과 타인의 저작권을 포함한 지적재산권 및 기타 권리를 침해하는 내용물에 대하여는 등록할 수 없으며, 만일 이와 같은 내용의 게시물로 인해 발생하는 결과에 대한 모든 책임은 회원 본인에게 있습니다.",
+        },
+        {
+          id: "4",
+          content:
+            "회사가 작성한 저작물에 대한 저작권 기타 지적재산권은 회사에 귀속합니다.",
+        },
+        {
+          id: "5",
+          content:
+            "회원은 서비스를 이용함으로써 얻은 정보를 회사의 사전 승낙 없이 복제, 송신, 출판, 전송, 배포, 방송 기타 방법에 의하여 영리목적으로 이용하거나 제3자에게 이용하게 하여서는 안됩니다.",
+        },
+        {
+          id: "6",
+          content:
+            "회원이 등록한 게시물에 대한 저작권은 해당 저작권자에게 귀속합니다.",
+        },
+      ],
+    },
+    {
+      id: "4",
+      title: "제4장 서비스 이용제한 및 계약해지",
+      content: "",
+    },
+    {
+      id: "4-1",
+      title: "제12조 서비스 이용제한",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 다음에 해당하는 경우 사전통지 없이 회원의 이용계약을 해지하거나 또는 일정기간 서비스 이용을 제한할 수 있습니다.",
+        },
+      ],
+      circledItems: [
+        {
+          id: "1",
+          content: "타인의 ID 및 비밀번호, 이메일 주소를 도용하는 경우",
+        },
+        {
+          id: "2",
+          content: "범죄행위와 관련되는 경우",
+        },
+        {
+          id: "3",
+          content: "공공질서 및 미풍양속에 위반는 경우",
+        },
+        {
+          id: "4",
+          content: "타인의 명예를 훼손하거나 불이익을 주는 경우",
+        },
+        {
+          id: "5",
+          content: "서비스에 위해를 가하는 등 건전한 이용을 저해하는 경우",
+        },
+        {
+          id: "6",
+          content: "기타 관계법령에 위배되는 경우",
+        },
+      ],
+    },
+    {
+      id: "4-2",
+      title: "제13조 계약의 해지",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회원이 이용계약을 해지하고자 하는 때에는 본인이 서비스 또는 E메일을 통하여 해지신청을 하여야 합니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 회원이 제13조의 내용을 위반하고, 회사 소정의 기간 내에에 이를 해소하지 아니하는 경우 서비스 이용계약을 해지 또는 일정기간 승낙을 유보할 수 있습니다.",
+        },
+      ],
+    },
+    {
+      id: "5",
+      title: "제5장 의 무",
+      content: "",
+    },
+    {
+      id: "5-1",
+      title: "제14조 회사의 의무",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 제9조 및 기타 특별한 사유가 없는 한 이용자가 신청한 서비스를 이용할 수 있도록 합니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 이 약관에서 정한 바에 따라 지속적, 안정적으로 서비스를 제공할 의무가 있습니다.",
+        },
+        {
+          id: "c",
+          content:
+            "회사는 회원의 개인정보를 본인의 승낙 없이 타인에게 공개, 배포하지 않습니다. 단, 전기통신관련 법령 등 관계법령에 의해 국가기관 등의 요구가 있는 경우, 개인의 안전을 보호해야 할 시급한 경우에는 그러하지 않습니다.",
+        },
+      ],
+    },
+    {
+      id: "5-2",
+      title: "제15조 회원의 의무",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content: "ID와 비밀번호에 관한 모든 관리의 책임은 회원에게 있습니다.",
+        },
+        {
+          id: "b",
+          content:
+            "ID와 비밀번호는 회사의 승낙 없이 다른 사람에게 양도, 대여할 수 없습니다.",
+        },
+        {
+          id: "c",
+          content:
+            "자신의 ID가 부정하게 사용된 경우 회원은 반드시 회사에 그 사실을 E메일이나 기타 방법을 이용하여 통보해야 합니다.",
+        },
+        {
+          id: "d",
+          content:
+            "회원은 이 약관 및 관계법령에서 규정한 사항을 준수해야 합니다.",
+        },
+        {
+          id: "e",
+          content:
+            "회원은 불법촬영물,허위영상물,아동성착취물로 판단되는 내용을 절대로 게재해서는 안되며 게재 시 관련법령에 의해 형사처벌을 받을 수도 있습니다.",
+        },
+      ],
+    },
+    {
+      id: "6",
+      title: "제6장 손해배상 및 면책사항",
+      content: "",
+    },
+    {
+      id: "6-1",
+      title: "제16조 손해배상",
+      content:
+        "회사는 무료로 제공되는 서비스와 관련하여 회원에게 손해가 발생하는 경우 이에 대하여 책임을 지지 않습니다.",
+    },
+    {
+      id: "6-2",
+      title: "제17조 면책사항",
+      content: "",
+      subItems: [
+        {
+          id: "a",
+          content:
+            "회사는 천재지변 및 이에 준하는 불가항력으로 인하여 서비스를 제공할 수 없는 경우에는 서비스 제공에 관한 책임이 면제됩니다.",
+        },
+        {
+          id: "b",
+          content:
+            "회사는 회원이 서비스에 게재한 정보 및 자료의 신뢰성, 정확성 등 내용에 관하여는 책임을 지지 않습니다.",
+        },
+        {
+          id: "c",
+          content:
+            "회원ID와 비밀번호 관리 및 이용상의 부주의로 인하여 발생되는 손해 또는 제3자에 의한 부정사용 등에 대한 모든 책임은 회원에게 있습니다.",
+        },
+        {
+          id: "d",
+          content:
+            "회원이 제13조 및 기타 약관에 규정된 사항을 위반하여 회사가 회원 또는 제3자에 대하여 책임을 부담하고 이로 인해 회사에 손해가 발생한 경우, 이 약관을 위반한 회원은 회사가 입은 모든 손해를 배상하여야 하며 동 손해로부터 회사를 면책시켜야 합니다.",
+        },
+      ],
+    },
+    {
+      id: "7",
+      title: "제18조 관할법원",
+      content:
+        "서비스 이용 중 발생한 회사와 이용자간의 분쟁에 대해 소송이 제기될 경우, 동 소송은 회사의 본사 소재지를 관할하는 법원을 관할법원으로 합니다.",
+    },
+  ],
+};
+
+export default termsServiceObject;
